Llama_Q_A_chatbot/requirements.txt
ipykernel
langchain_community
langchain
streamlit
langchain-chroma
langchain-core
python-dotenv
langchain_ollama
transformers